user signup with bctypt

+++ components/dashboard/OurProducts.jsx
@@ -1,33 +1,44 @@
 "use client";
 import { useContext, useEffect, useState } from "react";
 import Link from "next/link";
-
-import { FiEdit, FiTrash } from "react-icons/fi";
+import Skeleton from "react-loading-skeleton";
+import "react-loading-skeleton/dist/skeleton.css";
+import {
+  FiAlertTriangle,
+  FiCheckCircle,
+  FiEdit,
+  FiTrash,
+} from "react-icons/fi";
 import axios from "axios";
 import { AppContext } from "@/app/contextApi/contextProvider";
 import Image from "next/image";
 import EditProductModal from "./EditProductModal";
 import { useForm } from "react-hook-form";
+import ConfirmPopup from "../common/ConfirmPopup";
 
 const OurProducts = () => {
   // all products
-  const { products, setProducts } = useContext(AppContext);
+  const { products, setProducts, loading, setLoading } = useContext(AppContext);
   const [isModalOpen, setModalOpen] = useState(false);
   const [updateID, setUpdateID] = useState({});
   const [productsData, setProductsData] = useState({});
   const [currentPage, setCurrentPage] = useState(1);
   const { register, handleSubmit, reset } = useForm();
   const [selectedProducts, setSelectedProducts] = useState([]);
-
+  // confirm popup
+  const [isPopupOpen, setIsPopupOpen] = useState(false);
+  const [pendingDeleteId, setPendingDeleteId] = useState(null);
   const itemsPerPage = 10;
   // get all products
   const getProduct = async () => {
+    setLoading(true);
     try {
       const products = await axios.get(
         `http://localhost:5000/api/v1/product?page=${currentPage}&limit=${itemsPerPage}`
       );
       setProducts(products.data.products);
       setProductsData(products.data);
+      setLoading(false);
     } catch (error) {
       console.log(error.message);
     }
@@ -55,20 +66,29 @@ const OurProducts = () => {
       setCurrentPage(currentPage + 1);
     }
   }
-
+  //custom confirm delete
+  const handleConfirm = (id) => {
+    setIsPopupOpen(true);
+    setPendingDeleteId(id);
+  };
+  const handleCancel = () => setIsPopupOpen(!isPopupOpen);
   // delete a products
-  const deleteProduct = async (id) => {
-    try {
-      const response = await axios.delete(
-        `http://localhost:5000/api/v1/product/${id}`
-      );
-      if (response.status == 200) {
-        setProducts((prevProduct) =>
-          prevProduct.filter((product) => product._id !== id)
+  const deleteProduct = async () => {
+    if (pendingDeleteId) {
+      try {
+        const response = await axios.delete(
+          `http://localhost:5000/api/v1/product/${pendingDeleteId}`
         );
+        if (response.status == 200) {
+          setProducts((prevProduct) =>
+            prevProduct.filter((product) => product._id !== pendingDeleteId)
+          );
+          setIsPopupOpen(false);
+          setPendingDeleteId(null);
+        }
+      } catch (error) {
+        console.log(error.message);
       }
-    } catch (error) {
-      console.log(error.message);
     }
   };
 
@@ -83,13 +103,16 @@ const OurProducts = () => {
         );
 
         setProducts(response.data.slice(0, 10));
+        setLoading(false);
       } catch (error) {
         console.log(error.message);
       }
+    } else {
+      getProduct();
     }
   };
 
-  // modal open
+  // edit modal open
   const openModal = async (product) => {
     if (product._id) {
       setModalOpen(true);
@@ -102,10 +125,10 @@ const OurProducts = () => {
   };
   // Select all products
   const toggleSelectAll = () => {
-    if (selectedProducts.length === displayedProducts.length) {
+    if (selectedProducts.length === products.length) {
       setSelectedProducts([]);
     } else {
-      setSelectedProducts(displayedProducts.map((product) => product.id));
+      setSelectedProducts(products.map((product) => product._id));
     }
   };
 
@@ -118,13 +141,28 @@ const OurProducts = () => {
     );
   };
 
-  // Delete selected products
-  // const handleDelete = () => {
-  //   setProducts(
-  //     products.filter((product) => !selectedProducts.includes(product.id))
-  //   );
-  //   setSelectedProducts([]);
-  // };
+  // bulk delete product confirm
+  // const bulkDelete = ()=>{
+
+  // }
+  //bulk Delete selected products
+  const handleDelete = async () => {
+    try {
+      const response = await axios.delete(
+        `http://localhost:5000/api/v1/product/bulk-delete`,
+        {
+          data: {
+            id: { selectedProducts },
+          },
+        }
+      );
+
+      if (response.data.acknowledged) {
+        getProduct();
+        setSelectedProducts([]);
+      }
+    } catch (error) {}
+  };
   return (
     <div className="p-4 md:p-8 bg-gray-50 min-h-screen">
       <h1 className="text-2xl font-semibold text-gray-800 mb-6">
@@ -139,9 +177,12 @@ const OurProducts = () => {
           placeholder="Search products..."
           className="px-4 py-2 border rounded-md focus:outline-none focus:ring-2 focus:ring-indigo-500"
         />
+        <span className={`${selectedProducts.length > 0 && "text-red-700"}`}>
+          Selected {selectedProducts.length}
+        </span>
         <button
-          // onClick={handleDelete}
-          // disabled={!selectedProducts.length}
+          onClick={handleDelete}
+          disabled={!selectedProducts.length}
           className={`bg-red-600 text-white px-4 py-2 rounded-md hover:bg-red-700 transition duration-300 ${
             !selectedProducts.length ? "opacity-50 cursor-not-allowed" : ""
           }`}
@@ -152,60 +193,64 @@ const OurProducts = () => {
 
       {/* Product Table */}
       <div className="overflow-x-auto">
-        <table className="min-w-full bg-white border border-gray-200 rounded-lg shadow-md">
-          <thead>
-            <tr>
-              <th className="p-4 text-left">
-                <input
-                  type="checkbox"
-                  onChange={toggleSelectAll}
-                  // checked={selectedProducts.length === displayedProducts.length}
-                />
-              </th>
-              <th className="p-4 text-left">Product Image</th>
-              <th className="p-4 text-left">Product Name</th>
-              <th className="p-4 text-left">Price</th>
-              <th className="p-4 text-left">Stock</th>
-              <th className="p-4 text-left">Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            {products.map((product) => (
-              <tr key={product._id} className="hover:bg-gray-50">
-                <td className="p-4">
+        {loading ? (
+          <Skeleton height={50} count={6} />
+        ) : (
+          <table className="min-w-full bg-white border border-gray-200 rounded-lg shadow-md">
+            <thead>
+              <tr>
+                <th className="p-4 text-left">
                   <input
                     type="checkbox"
-                    checked={selectedProducts.includes(product.id)}
-                    onChange={() => toggleSelectProduct(product.id)}
+                    onChange={toggleSelectAll}
+                    checked={selectedProducts.length === products.length}
                   />
-                </td>
-                <td className="p-4">
-                  <div className="relative w-[50px] h-[50px] ">
-                    {" "}
-                    <Image src={product.image} alt="dairy milk" fill />
-                  </div>
-                </td>
-                <td className="p-4">{product.name}</td>
-                <td className="p-4">{product.price}</td>
-                <td className="p-4">{product.stock}</td>
-                <td className="p-4 flex items-center gap-2">
-                  <button
-                    onClick={() => openModal(product)}
-                    className="text-indigo-600 hover:text-indigo-800"
-                  >
-                    <FiEdit />
-                  </button>
-                  <button
-                    onClick={() => deleteProduct(product._id)}
-                    className="text-red-600 hover:text-red-800"
-                  >
-                    <FiTrash />
-                  </button>
-                </td>
+                </th>
+                <th className="p-4 text-left">Product Image</th>
+                <th className="p-4 text-left">Product Name</th>
+                <th className="p-4 text-left">Price</th>
+                <th className="p-4 text-left">Stock</th>
+                <th className="p-4 text-left">Actions</th>
               </tr>
-            ))}
-          </tbody>
-        </table>
+            </thead>
+            <tbody>
+              {products.map((product) => (
+                <tr key={product._id} className="hover:bg-gray-50">
+                  <td className="p-4">
+                    <input
+                      type="checkbox"
+                      checked={selectedProducts.includes(product._id)}
+                      onChange={() => toggleSelectProduct(product._id)}
+                    />
+                  </td>
+                  <td className="p-4">
+                    <div className="relative w-[50px] h-[50px] ">
+                      {" "}
+                      <Image src={product?.image} alt="dairy milk" fill />
+                    </div>
+                  </td>
+                  <td className="p-4">{product.name}</td>
+                  <td className="p-4">{product.price}</td>
+                  <td className="p-4">{product.stock}</td>
+                  <td className="p-4 flex items-center gap-2">
+                    <button
+                      onClick={() => openModal(product)}
+                      className="text-indigo-600 hover:text-indigo-800"
+                    >
+                      <FiEdit />
+                    </button>
+                    <button
+                      onClick={() => handleConfirm(product._id)}
+                      className="text-red-600 hover:text-red-800"
+                    >
+                      <FiTrash />
+                    </button>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        )}
       </div>
 
       {/* Pagination */}
@@ -218,7 +263,8 @@ const OurProducts = () => {
           Previous
         </button>
         <span className="text-gray-600">
-          Page {currentPage}/{productsData.totalPages}
+          Page {currentPage}/{productsData.totalPages}, Products{" "}
+          {`${products.length}`}
         </span>
         <button
           onClick={handleNext}
@@ -228,6 +274,16 @@ const OurProducts = () => {
           Next
         </button>
       </div>
+
+      {/* tost test */}
+      <ConfirmPopup
+        message="Are you sure you want to delete this item?"
+        onConfirm={handleConfirm}
+        onCancel={handleCancel}
+        isOpen={isPopupOpen}
+        deleteProduct={deleteProduct}
+      />
+
       <EditProductModal
         isOpen={isModalOpen}
         openModal={openModal}
